requirements.txt
Django==1.9.5
Django-Select2==5.8.4
Twisted==16.1.1
Werkzeug==0.11.5
asgiref==0.11.2
autobahn==0.13.1
boto==2.39.0
channels==0.11.0
click==6.6
daphne==0.10.3
django-advanced-filters==1.0.4
django-appconf==1.0.1
django-autocomplete-light==3.1.5
django-bower==5.1.0
django-braces==1.4.0
django-easy-select2==1.2.5
django-extensions==1.6.1
django-jsonfield==0.9.19
django-pipeline==1.6.8
django-querysetsequence==0.5
django-rq==0.9.0
django-s3direct==0.3.12
django-storages==1.4.1
django-taggit==0.18.1
honcho==0.6.6
redis==2.10.5
rq==0.5.6
simplejson==3.6.5
six==1.10.0
txaio==2.3.1
wsgiref==0.1.2
zope.interface==4.1.3
